+++ main.py
@@ -1,33 +1,46 @@
+# ==== GAME MADE BY ====
+# ====   CRAFTEDGP  ====
+# ====     A.K.A    ====
+# ====  AETHERFLARE ====
+# ====     GAMES    ====
+
 from ursina import *
 from direct.actor.Actor import Actor
 from ursina.shaders import lit_with_shadows_shader, fxaa_shader, ssao_shader
 
-# ======= <CONFIG> =======
-
-# ======= </CONFIG> =======
-
 app = Ursina(title='Flux Takedown', fullscreen=True, borderless=False, show_ursina_splash=True)
 
 # ======= <Top Down Controller> =======
-mouse.visible = False
-idle_animation = Actor('assets/animations/player/idle_combat.gltf')
+# ---- <Animations> ----
+idle_animation = Actor('assets/animations/player/idle_gun.gltf')
 idle_animation.loop('Armature|mixamo.com|Layer0')
-sprint_actor = Actor('assets/animations/player/sprint.gltf')
-sprint_actor.loop('Armature|mixamo.com|Layer0')
+sprint_animation = Actor('assets/animations/player/sprint.gltf')
+sprint_animation.loop('Armature|mixamo.com|Layer0')
+falling_unarmed_animation = Actor('assets/animations/player/falling_unarmed.gltf')
+falling_unarmed_animation.loop('Armature|mixamo.com|Layer0')
+# <Correct broken animations> 
+idle_animation.setHpr(180, 0, 0)
+sprint_animation.setHpr(180, 0, 0)
+falling_unarmed_animation.setHpr(180, 0, 0)
+# </Correct broken animations>
+# ---- </Animations> ----
+mouse.visible = False
 player = Entity(model=idle_animation, collider='box', position=(0, 100, 0))
 crosshair = Entity(model='quad', texture='assets/gfx/crosshair.png', parent=camera.ui, scale=(.1, .1), position=(0, 0, 2), visible=True)
 camera.rotation_x = 90
 in_game = True
+time_interval = 0.01
+previous_position = player.position
 # ======= </Top Down Controller> ======
 
 # ======= <Raycast Collision Handling and Gravity> =======
 def move_with_raycast(direction):
-    speed = 8 * time.dt
+    speed = 3 * time.dt
     ray = raycast(player.world_position, 
                   direction, 
                   ignore=(player,), 
                   distance=0.5, debug=True)
-    
+
     if not ray.hit:
         player.position += direction * speed
 
@@ -38,6 +51,7 @@ max_fall_speed = 100
 
 def apply_gravity():
     global fall_speed
+    global is_grounded
     gravity_ray = raycast(player.world_position, 
                           Vec3(0, -1, 0), 
                           ignore=(player,), 
@@ -45,16 +59,24 @@ def apply_gravity():
                           debug=True)
 
     if not gravity_ray.hit:
+        is_grounded = False
         fall_speed = min(fall_speed + gravity_force * time.dt, max_fall_speed)
-        below_player_ray = raycast(player.world_position + Vec3(0, -0.5, 0), Vec3(0, -1, 0), ignore=(player,), distance=fall_speed * time.dt, debug=True)
+        below_player_ray = raycast(player.world_position + Vec3(0, -0.5, 0), 
+                                   Vec3(0, -1, 0), 
+                                   ignore=(player,), 
+                                   distance=0.6, 
+                                   debug=True)
+        
         if not below_player_ray.hit:
             player.y -= fall_speed * time.dt
     else:
         fall_speed = 0
+        is_grounded = True
+
 # ======= </Raycast Collision Handling and Gravity> ========
 
 # ======= <MAP LAYOUT> =======
-camera.shader = fxaa_shader
+# camera.shader = fxaa_shader
 # camera.shader = ssao_shader
 Sky()
 Entity.default_shader = lit_with_shadows_shader
@@ -75,10 +97,10 @@ gun = Entity(model='assets/items/gun/gun.obj', position=(.1, 0, .5), scale=0.1, 
 bullets = []
 def shoot():
     bullet = Entity(model='sphere', color=color.yellow, scale=0.2, position=gun.world_position, collider=None)
-    bullet.direction = gun.forward  
-    bullet.speed = 100
-    bullet.previous_position = bullet.position 
-    bullets.append(bullet) 
+    bullet.direction = gun.forward
+    bullet.speed = 300
+    bullet.previous_position = bullet.position
+    bullets.append(bullet)
 # ======= </HANDLING SHOOTING LOGIC> =======
 
 # ======= <UI> =======
@@ -89,8 +111,8 @@ frame = Entity(model='wireframe_cube', color=color.white33, position=(-.47, -.03
 walk_controls = Text('W, A, S, D', position=(-.47, .1), origin=(0, 0), font=controls_font)
 rightclick_control = Text('Right Click : Pickup/Throw weapons', position=(-.47, 0), origin=(0, 0), font=controls_font, scale=0.9)
 leftclick_control = Text('Left Click : Shoot/Melee', position=(-.47, -.1), origin=(0, 0), font=controls_font)
-swipe_control = Text('Swipe Mouse : Curve bullets', position=(-.47, -.2), origin=(0, 0), font=controls_font)
-controls_list = [title_controls, frame, walk_controls, rightclick_control, leftclick_control, swipe_control]
+# swipe_control = Text('Swipe Mouse : Curve bullets', position=(-.47, -.2), origin=(0, 0), font=controls_font)
+controls_list = [title_controls, frame, walk_controls, rightclick_control, leftclick_control]
 for i in controls_list:
     invoke(i.fade_out(duration=10))
 # ---- </Start Screen> ----
@@ -117,7 +139,7 @@ def options():
     placeholder = Text('Comming soon', color=color.red, position=(0, .2), origin=(0, 0), scale=1.5, font='assets/fonts/Exo-Regular.ttf')
     btn_back = Button(text='Back', position=(0, 0), origin=(0, 0), scale=(0.2, 0.05), on_click=back)
 # ---- </Button Commands> ----
-version = Text('Flux Takedown Alpha-0.0.6', color=color.white, position=window.top_left, scale=1)
+version = Text('Flux Takedown Alpha-0.0.7', color=color.white, position=window.top_left, scale=1)
 bg = Panel(origin=(0, 0), position=(0, 0), alpha=0.5, visible=False)
 bg.scale_x = camera.aspect_ratio
 bg.scale_y = camera.aspect_ratio
@@ -132,30 +154,27 @@ def update():
     # <Deagle>
     if in_game:
         for bullet in bullets:
-            bullet.previous_position = bullet.position 
+            bullet.previous_position = bullet.position
             bullet.position += bullet.direction * bullet.speed * time.dt
 
             ray = raycast(bullet.previous_position, (bullet.position - bullet.previous_position).normalized(), ignore=(bullet,), distance=.3, debug=True)
 
-            if ray.hit: 
-                destroy(bullet)  
-                bullets.remove(bullet)  
+            if ray.hit:
+                destroy(bullet)
+                bullets.remove(bullet)
                 return
 
             if distance(player.position, bullet.position) > 40:
-                destroy(bullet) 
+                destroy(bullet)
                 bullets.remove(bullet)
     # </Deagle>
     # ---- </Shooting> ----
     # ---- <Gravity> ----
-    apply_gravity()
+    if in_game:
+        apply_gravity()
     # ---- </Gravity> ----
-    # ---- <Correct broken animations> ----
-    idle_animation.setHpr(180, 0, 0)
-    sprint_actor.setHpr(180, 0, 0)
-    # ---- </Correct broken animations> ----
     # ---- <Update Crosshair and Player Position/Roation> ----
-    camera.position = lerp(camera.position, (player.x, player.y + 20, player.z), 10 * time.dt)
+    camera.position = lerp(camera.position, (player.x, player.y + 15, player.z), 10 * time.dt)
     crosshair.position = mouse.position
     if in_game:
         player.look_at_xz(Vec3(mouse.x + player.x, 0, mouse.y + player.z))
@@ -173,7 +192,26 @@ def update():
 
         if held_keys['d']:
             move_with_raycast(Vec3(1, 0, 0))
+
     # ---- </Controls> ----
+    # ---- <Animation Handling> ----
+    global previous_position
+    global is_grounded
+
+    if is_grounded:
+        if player.position != previous_position:
+            player.model = sprint_animation
+        else:  
+            player.model = idle_animation
+    else:
+        player.model = falling_unarmed_animation
+
+    invoke(lambda: store_position(), delay=time_interval)
+
+    def store_position():
+        global previous_position
+        previous_position = player.position
+    # ---- </Animation Handling > ----
 
 def input(key):
     # ---- <Making variables global if needed, always do it here before anything> ----
@@ -217,17 +255,17 @@ def input(key):
     # ---- </Pause Menu> ----
 
     # ---- <Animation handling> ----
-    if in_game:
-        if held_keys['w']:
-            player.model = sprint_actor
-        elif held_keys['s']:
-            player.model = sprint_actor
-        elif held_keys['a']:
-            player.model = sprint_actor
-        elif held_keys['d']:
-            player.model = sprint_actor
-        else:
-            player.model = idle_animation
+    # if in_game:
+        # if held_keys['w']:
+            # player.model = sprint_animation
+        # elif held_keys['s']:
+            # player.model = sprint_animation
+        # elif held_keys['a']:
+            # player.model = sprint_animation
+        # elif held_keys['d']:
+            # player.model = sprint_animation
+        # else:
+            # player.model = idle_animation
     # ---- </Animation handling> ----
 # ======= </UPDATE AND INPUT FUNCTION FOR EVERYTHING> ======
 app.run()